New fixture to create Gradle Project for AGP integ testing
The new GradleRule] handles project creation, which is configured
via the recently added GradleBuild/GradleProject.

The rule then exposes Build/SubProject to the test to do limited
manipulation and run builds (or syncs).

This also wires the project configuration to the recently added
DslProxy class over the normal AGP extension, allowing all
possibly DSL to be used in the fixture.

Finally, a first test (DynamicFeatureNamespaceTest) is converted
to the new fixture.

Bug: N/A
Test: Unit tests where applicable.
Change-Id: Ib98ab9e9d4486ca5f0b210ae9779b6e98dd6a0e6

Former-commit-id: 435064a0d78c87e1664bef0193438db29776215e