[GMD] surface emulator error messages for setup
Surface emulator kernel error messages when the emulator closes
unexpectedly.

The emulator has kernel messages that are useful for determining the
root cause of the avd/snapshot creation failure. It will be useful to
surface these to users; not only in case that they might be able to
remedy the situation immediately, but also help us identify failures
more quickly

Bug: 331016661
Test: Added

Change-Id: If9303e70c263f04699967bf19baad834d8156c79

Former-commit-id: 0844d8cc8ab4db3a58b897cf1ff1af11d6a87648